requirements.txt
discord.py
python-dotenv
pyinstaller
